Move writing addon data to its own file
