+++ src/components/LandingPage.vue
@@ -259,11 +259,10 @@
 </template>
 
 <script lang="ts">
-import backupIfRequired from "@/libs/backup";
 import contacts from "@/libs/contacts";
 import { grabVersionFromToc } from "@/libs/grab-wa-version";
 import hash from "@/libs/hash";
-import sanitize from "@/libs/sanitize";
+import { writeAddonData } from "@/libs/write-addon-data";
 import {
   createSortByAuthor,
   createSortByString,
@@ -469,7 +468,7 @@ export default defineComponent({
   watch: {
     "stash.auras": {
       handler() {
-        this.writeAddonData();
+        writeAddonData(this.config, this.addonsInstalled, this.aurasWithData, this.stash);
       },
       deep: true,
     },
@@ -1147,7 +1146,7 @@ export default defineComponent({
         // No data for any addon available. Nothing to update.
         try {
           if (this.IsAddonInstalled("WeakAuras") || this.IsAddonInstalled("Plater")) {
-            this.writeAddonData();
+            writeAddonData(this.config, this.addonsInstalled, this.aurasWithData, this.stash);
           }
         } finally {
           this.fetching = false;
@@ -1240,7 +1239,7 @@ export default defineComponent({
               // We are done with the Wago API, update data.lua
 
               try {
-                this.writeAddonData();
+                writeAddonData(this.config, this.addonsInstalled, this.aurasWithData, this.stash);
               } finally {
                 this.fetching = false;
 
@@ -1269,284 +1268,8 @@ export default defineComponent({
     toggleUpdatedAuraList() {
       this.updateAuraIsShown = !this.updateAuraIsShown;
     },
-    async writeAddonData() {
-      console.log("writeAddonData");
-      const addonConfigs = this.addonsInstalled;
-
-      if (this.config.wowpath.validated && this.config.wowpath.version !== "") {
-        var AddonPath = ["Interface", "AddOns", "WeakAurasCompanion"];
-        var AddonFolder = path.join(this.config.wowpath.value, this.config.wowpath.version);
-
-        while (AddonPath.length) {
-          var check = AddonPath.shift();
-          var folder = matchFolderNameInsensitive(AddonFolder, check, AddonPath.length === 0);
-
-          if (folder) {
-            AddonFolder = path.join(AddonFolder, folder);
-          } else {
-            throw new Error("errorCantCreateAddon");
-          }
-        }
-        // Make data.lua
-        let LuaOutput = "-- file generated automatically\n";
-        LuaOutput += "WeakAurasCompanionData = {\n";
-        let addonDepts = "";
-        const fields = [
-          "name",
-          "author",
-          "encoded",
-          "wagoVersion",
-          "wagoSemver",
-          "source",
-          // "logo" -- keep that for a future WeakAuras release
-        ];
-
-        addonConfigs.forEach((config) => {
-          addonDepts += config.addonName + ",";
-
-          let spacing = "";
-
-          LuaOutput += `  ${config.addonName} = {\n`;
-          spacing = "  ";
-
-          let LuaSlugs = spacing + "  slugs = {\n";
-
-          this.aurasWithData
-            .filter((aura) => aura.auraType === config.addonName)
-            .forEach((aura) => {
-              LuaSlugs += spacing + `    ["${aura.slug.replace(/"/g, '\\"')}"] = {\n`;
-
-              fields.forEach((field) => {
-                if (aura[field]) {
-                  LuaSlugs += spacing + `      ${field} = [=[${aura[field]}]=],\n`;
-                }
-              });
-
-              if (typeof aura.changelog !== "undefined") {
-                if (typeof aura.changelog.text !== "undefined") {
-                  let sanitized;
-
-                  if (aura.changelog.format === "bbcode") {
-                    sanitized = sanitize.bbcode(aura.changelog.text);
-                  } else if (aura.changelog.format === "markdown") {
-                    sanitized = sanitize.markdown(aura.changelog.text);
-                  }
-
-                  LuaSlugs += spacing + `      versionNote = [=[${sanitized}]=],\n`;
-                }
-              }
-
-              LuaSlugs += spacing + "    },\n";
-            });
-          LuaOutput += LuaSlugs;
-          LuaOutput += spacing + "  },\n";
-          LuaOutput += spacing + "  stash = {\n";
-
-          this.stash.auras
-            .filter((aura) => aura.auraType === config.addonName)
-            .forEach((aura) => {
-              LuaOutput += spacing + `    ["${aura.slug.replace(/"/g, '\\"')}"] = {\n`;
-
-              fields.forEach((field) => {
-                if (aura[field]) {
-                  LuaOutput += spacing + `      ${field} = [=[${aura[field]}]=],\n`;
-                }
-              });
-
-              if (typeof aura.changelog !== "undefined") {
-                if (typeof aura.changelog.text !== "undefined") {
-                  let sanitized;
-
-                  if (aura.changelog.format === "bbcode") {
-                    sanitized = sanitize.bbcode(aura.changelog.text);
-                  } else if (aura.changelog.format === "markdown") {
-                    sanitized = sanitize.markdown(aura.changelog.text);
-                  }
-
-                  LuaOutput += spacing + `      versionNote = [=[${sanitized}]=],\n`;
-                }
-              }
-
-              LuaOutput += spacing + `      source = "${aura.source}",\n`;
-              LuaOutput += spacing + "    },\n";
-            });
-
-          LuaOutput += spacing + "  },\n";
-
-          if (config.addonName === "WeakAuras") {
-            LuaOutput += spacing + "  stopmotionFiles = {\n";
-            const stopmotionFilesPath = path.join(AddonFolder, "animations");
-
-            if (fs.existsSync(stopmotionFilesPath)) {
-              const regex = new RegExp(/^(.*?)(?: GIF)?\.x\d+y\d+f\d+w\d+h\d+W\d+H\d+\.tga$/);
-
-              fs.readdirSync(stopmotionFilesPath)
-                .filter((v) => v && v.match(regex))
-                .map((v) => ({
-                  filename: v,
-                  title: v.match(regex)[1],
-                }))
-                .forEach((file) => {
-                  LuaOutput += spacing + `    [ [=[${file.filename}]=] ] = [=[${file.title}]=],\n`;
-                });
-            }
-
-            LuaOutput += spacing + "  },\n";
-          }
-          LuaOutput += "  },\n";
-        });
-
-        LuaOutput += "}";
-
-        /* if (this.stash.lenghth > 0) { LuaOutput += "" } */
-        const tocVersion = this.grabWATocVersion(this.config.wowpath.value, this.config.wowpath.version);
-        console.log("WeakAuras.toc has version:", tocVersion);
-        const files = [
-          {
-            name: "WeakAurasCompanion.toc",
-            data: `## Interface: ${tocVersion}
-## Title: WeakAuras Companion
-## Author: The WeakAuras Team
-## Version: ${__APP_VERSION__}
-## IconTexture: Interface\\AddOns\\WeakAuras\\Media\\Textures\\icon.blp
-## Notes: Keep your WeakAuras updated!
-## X-Category: Interface Enhancements
-## DefaultState: Enabled
-## LoadOnDemand: 0
-## OptionalDeps: ${addonDepts}
-
-data.lua
-init.lua`,
-          },
-          {
-            name: "WeakAurasCompanion-Classic.toc",
-            data: `## Interface: ${tocVersion}
-## Title: WeakAuras Companion
-## Author: The WeakAuras Team
-## Version: ${__APP_VERSION__}
-## Notes: Keep your WeakAuras updated!
-## X-Category: Interface Enhancements
-## DefaultState: Enabled
-## LoadOnDemand: 0
-## OptionalDeps: ${addonDepts}
-
-data.lua
-init.lua`,
-          },
-          {
-            name: "WeakAurasCompanion-Mainline.toc",
-            data: `## Interface: ${tocVersion}
-## Title: WeakAuras Companion
-## Author: The WeakAuras Team
-## Version: ${__APP_VERSION__}
-## IconTexture: Interface\\AddOns\\WeakAuras\\Media\\Textures\\icon.blp
-## Notes: Keep your WeakAuras updated!
-## X-Category: Interface Enhancements
-## DefaultState: Enabled
-## LoadOnDemand: 0
-## OptionalDeps: ${addonDepts}
-
-data.lua
-init.lua`,
-          },
-          {
-            name: "WeakAurasCompanion-Wrath.toc",
-            data: `## Interface: ${tocVersion}
-## Title: WeakAuras Companion
-## Author: The WeakAuras Team
-## Version: ${__APP_VERSION__}
-## Notes: Keep your WeakAuras updated!
-## X-Category: Interface Enhancements
-## DefaultState: Enabled
-## LoadOnDemand: 0
-## OptionalDeps: ${addonDepts}
-
-data.lua
-init.lua`,
-          },
-          {
-            name: "init.lua",
-            data: `-- file generated automatically
-local loadedFrame = CreateFrame("FRAME")
-loadedFrame:RegisterEvent("ADDON_LOADED")
-loadedFrame:SetScript("OnEvent", function(_, _, addonName)
-  if addonName == "WeakAurasCompanion" then
-    if WeakAuras and WeakAuras.AddCompanionData and WeakAurasCompanionData then
-      local WeakAurasData = WeakAurasCompanionData.WeakAuras
-      if WeakAurasData then
-        WeakAuras.AddCompanionData(WeakAurasData)
-        WeakAuras.StopMotion.texture_types["WeakAuras Companion"] = WeakAuras.StopMotion.texture_types["WeakAuras Companion"] or {}
-        local CompanionTextures = WeakAuras.StopMotion.texture_types["WeakAuras Companion"]
-        for fileName, name in pairs(WeakAurasData.stopmotionFiles) do
-          CompanionTextures["Interface\\\\AddOns\\\\WeakAurasCompanion\\\\animations\\\\" .. fileName] = name
-        end
-      end
-    end
-
-    if Plater and Plater.AddCompanionData and WeakAurasCompanionData and WeakAurasCompanionData.Plater then
-      Plater.AddCompanionData(WeakAurasCompanionData.Plater)
-    end
-  end
-end)
-`,
-          },
-          {
-            name: "data.lua",
-            data: LuaOutput,
-          },
-        ];
-
-        files.forEach((file) => {
-          let filepath = path.join(AddonFolder, file.name);
-
-          fs.writeFile(filepath, file.data, (err2) => {
-            if (err2) {
-              throw new Error("errorFileSave");
-            }
-          });
-        });
-      }
-      this.backup();
-    },
     installUpdates() {
       ipcRenderer.invoke("installUpdates");
-    },
-    backup() {
-      this.config.wowpath.versions.forEach((version) => {
-        version.accounts.forEach((account) => {
-          this.addonsInstalled.forEach((addon) => {
-            let lastSavedFileSize = null;
-
-            if (typeof account.savedvariableSizeForAddon === "undefined") account.savedvariableSizeForAddon = [];
-
-            const savedData = account.savedvariableSizeForAddon.find(
-              (savedAddon) => savedAddon.addonName === addon.addonName
-            );
-
-            if (savedData) {
-              lastSavedFileSize = savedData.fileSize;
-            }
-
-            backupIfRequired(
-              addon.svPathFunction(version.name, account.name),
-              this.config.backup,
-              lastSavedFileSize,
-              `${version.name}#${account.name}`,
-              (fileSize) => {
-                if (savedData) {
-                  savedData.fileSize = fileSize;
-                } else {
-                  account.savedvariableSizeForAddon.push({
-                    fileSize: fileSize,
-                    addonName: addon.addonName,
-                  });
-                }
-              },
-              addon.addonName
-            );
-          });
-        });
-      });
     },
     validateWowpath() {
       console.log("validateWowpath");
